feat: 룸 페이지에서 Building 컴포넌트 분리
